+++ src/NEW/checks/notNullable.ts
@@ -1,8 +1,0 @@
-import { PredicateState } from '../types';
-
-/**
- * Disallow null values.
- */
-export default function nonNullable<T>(state: PredicateState<T>) {
-  state.nullable = false;
-}