+++ src/NEW/checks/notRequired.ts
@@ -1,8 +1,0 @@
-import { PredicateState } from '../types';
-
-/**
- * Require this field to NOT be explicitly defined.
- */
-export default function nonRequired<T>(state: PredicateState<T>) {
-  state.required = false;
-}