+++ src/NEW/checks/nullable.ts
@@ -1,8 +1,0 @@
-import { PredicateState } from '../types';
-
-/**
- * Allow null values.
- */
-export default function nullable<T>(state: PredicateState<T>) {
-  state.nullable = true;
-}