+++ src/NEW/checks/required.ts
@@ -1,8 +1,0 @@
-import { PredicateState } from '../types';
-
-/**
- * Require this field to be explicitly defined.
- */
-export default function required<T>(state: PredicateState<T>) {
-  state.required = true;
-}