+++ src/NEW/criteria/custom.ts
@@ -1,13 +1,13 @@
 import { invariant } from '../helpers';
-import { CheckerCallback, CustomCallback, PredicateState } from '../types';
+import { CriteriaValidator, CustomCallback, SchemaState } from '../types';
 
 /**
  * Set a callback to run custom validation logic.
  */
 export default function custom<T>(
-  state: PredicateState<T>,
+  state: SchemaState<T>,
   callback: CustomCallback<T>,
-): void | CheckerCallback<T> {
+): void | CriteriaValidator<T> {
   if (__DEV__) {
     invariant(typeof callback === 'function', 'Custom requires a validation function.');