+++ src/NEW/criteria/notNullable.ts
@@ -1,0 +1,8 @@
+import { SchemaState } from '../types';
+
+/**
+ * Disallow null values.
+ */
+export default function nonNullable<T>(state: SchemaState<T>) {
+  state.nullable = false;
+}