+++ src/NEW/criteria/nullable.ts
@@ -1,0 +1,8 @@
+import { SchemaState } from '../types';
+
+/**
+ * Allow null values.
+ */
+export default function nullable<T>(state: SchemaState<T>) {
+  state.nullable = true;
+}