+++ src/NEW/criteria/required.ts
@@ -1,0 +1,8 @@
+import { SchemaState } from '../types';
+
+/**
+ * Require this field to be explicitly defined.
+ */
+export default function required<T>(state: SchemaState<T>) {
+  state.required = true;
+}